Cast strtol return to unsigned long
	* testsuite/test-strtol.c (run_tests): Cast strtol return to
	unsigned long.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@217429 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libiberty/testsuite/test-strtol.c
@@ -132,7 +132,8 @@ run_tests (const struct test_data_t *test_data, size_t ntests)
       switch (test_data[i].fun)
 	{
 	case STRTOL:
-	  res = strtol (test_data[i].nptr, 0, test_data[i].base);
+	  res = (unsigned long) strtol (test_data[i].nptr,
+					0, test_data[i].base);
 	  break;
 	case STRTOUL:
 	  res = strtoul (test_data[i].nptr, 0, test_data[i].base);
